requirements.txt
Flask
gunicorn
Jinja2
MarkupSafe
Werkzeug
numpy
scipy
nltk
scikit-learn==1.3.2
pandas
beautifulsoup4
jsonschema
tmdbv3api
lxml
urllib3
requests
pickleshare
python-dotenv
